Move Counter class into MyQue.
